Update index page to remove redirected www links
A few links from the Workshop Overview section still point to
www.datacarpentry.org. The domain being used now doesn't
include the www. This updates the old links in the page to match
this usage.

This doesn't cause any issues with the website, but it
is causing us some issues in a project we're working on to
help facilitation computational training where internet access
is limited.